Initial implementation.

+++ components/director/internal/domain/application/service.go
@@ -31,7 +31,6 @@ import (
 	"github.com/kyma-incubator/compass/components/director/internal/labelfilter"
 	"github.com/kyma-incubator/compass/components/director/internal/model"
 	"github.com/kyma-incubator/compass/components/director/internal/timestamp"
-	"github.com/kyma-incubator/compass/components/director/pkg/pagination"
 	"github.com/pkg/errors"
 )
 
@@ -120,6 +119,7 @@ type WebhookRepository interface {
 type FormationService interface {
 	AssignFormation(ctx context.Context, tnt, objectID string, objectType graphql.FormationObjectType, formation model.Formation) (*model.Formation, error)
 	UnassignFormation(ctx context.Context, tnt, objectID string, objectType graphql.FormationObjectType, formation model.Formation) (*model.Formation, error)
+	ListFormationsForObject(ctx context.Context, objectID string) ([]*model.Formation, error)
 }
 
 // RuntimeRepository missing godoc
@@ -280,32 +280,14 @@ func (s *service) ListByRuntimeID(ctx context.Context, runtimeID uuid.UUID, page
 		return nil, apperrors.NewInvalidDataError("runtime does not exist")
 	}
 
-	scenariosLabel, err := s.labelRepo.GetByKey(ctx, tenantID, model.RuntimeLabelableObject, runtimeID.String(), model.ScenariosKey)
+	formations, err := s.formationService.ListFormationsForObject(ctx, runtimeID.String())
 	if err != nil {
-		if apperrors.IsNotFoundError(err) {
-			return &model.ApplicationPage{
-				Data:       []*model.Application{},
-				PageInfo:   &pagination.Page{},
-				TotalCount: 0,
-			}, nil
-		}
-		return nil, errors.Wrap(err, "while getting scenarios for runtime")
+		return nil, errors.Wrapf(err, "while listing formations for runtime with ID %s", runtimeID.String())
 	}
 
-	scenarios, err := label.ValueToStringsSlice(scenariosLabel.Value)
-	if err != nil {
-		return nil, errors.Wrap(err, "while converting scenarios labels")
-	}
-	if len(scenarios) == 0 {
-		return &model.ApplicationPage{
-			Data:       []*model.Application{},
-			TotalCount: 0,
-			PageInfo: &pagination.Page{
-				StartCursor: "",
-				EndCursor:   "",
-				HasNextPage: false,
-			},
-		}, nil
+	formationNames := make([]string, 0, len(formations))
+	for _, formation := range formations {
+		formationNames = append(formationNames, formation.Name)
 	}
 
 	hidingSelectors, err := s.appHideCfgProvider.GetApplicationHideSelectors()
@@ -313,7 +295,7 @@ func (s *service) ListByRuntimeID(ctx context.Context, runtimeID uuid.UUID, page
 		return nil, errors.Wrap(err, "while getting application hide selectors from config")
 	}
 
-	return s.appRepo.ListByScenarios(ctx, tenantUUID, scenarios, pageSize, cursor, hidingSelectors)
+	return s.appRepo.ListByScenarios(ctx, tenantUUID, formationNames, pageSize, cursor, hidingSelectors)
 }
 
 // Get missing godoc
@@ -786,6 +768,10 @@ func (s *service) SetLabel(ctx context.Context, labelInput *model.LabelInput) er
 		return apperrors.NewNotFoundError(resource.Application, labelInput.ObjectID)
 	}
 
+	if labelInput.Key == model.ScenariosKey {
+		return errors.Errorf("label with key %s cannot be set explicitly", model.ScenariosKey)
+	}
+
 	err = s.labelService.UpsertLabel(ctx, appTenant, labelInput)
 	if err != nil {
 		return errors.Wrapf(err, "while creating label for Application")
@@ -815,39 +801,6 @@ func (s *service) GetLabel(ctx context.Context, applicationID string, key string
 	}
 
 	return label, nil
-}
-
-// GetScenariosGlobal list the scenario labels for the application globally and merges their values
-func (s *service) GetScenariosGlobal(ctx context.Context, applicationID string) ([]string, error) {
-	appTenant, err := tenant.LoadFromContext(ctx)
-	if err != nil {
-		return nil, errors.Wrapf(err, "while loading tenant from context")
-	}
-
-	appExists, err := s.appRepo.Exists(ctx, appTenant, applicationID)
-	if err != nil {
-		return nil, errors.Wrapf(err, "while checking the existence of Application with ID: %s", applicationID)
-	}
-	if !appExists {
-		return nil, fmt.Errorf("application with ID %s doesn't exist", applicationID)
-	}
-
-	labels, err := s.labelRepo.ListGlobalByKeyAndObjects(ctx, model.ApplicationLabelableObject, []string{applicationID}, model.ScenariosKey)
-	if err != nil {
-		return nil, errors.Wrapf(err, "while getting label for Application with ID: %s", applicationID)
-	}
-
-	var scenarios []string
-	for _, lbl := range labels {
-		scenariosFromLabel, err := label.ValueToStringsSlice(lbl.Value)
-		if err != nil {
-			return nil, errors.Wrapf(err, "while parsing label values for Application with ID: %s", applicationID)
-		}
-
-		scenarios = append(scenarios, scenariosFromLabel...)
-	}
-
-	return scenarios, nil
 }
 
 // ListLabels missing godoc
@@ -906,6 +859,10 @@ func (s *service) DeleteLabel(ctx context.Context, applicationID string, key str
 	}
 	if !appExists {
 		return errors.Errorf("application with ID %s doesn't exist", applicationID)
+	}
+
+	if key == model.ScenariosKey {
+		return errors.Errorf("label with key %s cannot be deleted explicitly", model.ScenariosKey)
 	}
 
 	err = s.labelRepo.Delete(ctx, appTenant, model.ApplicationLabelableObject, applicationID, key)
@@ -1069,13 +1026,18 @@ func (s *service) handleMergeLabels(ctx context.Context, srcAppLabels, destAppLa
 
 // ensureApplicationNotPartOfScenarioWithRuntime Checks if an application has scenarios associated with it. if a runtime is part of any scenario, then the application is considered being used by that runtime.
 func (s *service) ensureApplicationNotPartOfScenarioWithRuntime(ctx context.Context, tenant, appID string) error {
-	scenarios, err := s.getScenarioNamesForApplication(ctx, appID)
+	formations, err := s.formationService.ListFormationsForObject(ctx, appID)
 	if err != nil {
-		return err
+		return errors.Wrapf(err, "while getting formations for Application with ID %s", appID)
 	}
 
-	if len(scenarios) > 0 {
-		runtimes, err := s.getRuntimeNamesForScenarios(ctx, tenant, scenarios)
+	if len(formations) > 0 {
+		formationNames := make([]string, 0, len(formations))
+		for _, formation := range formations {
+			formationNames = append(formationNames, formation.Name)
+		}
+
+		runtimes, err := s.getRuntimeNamesForScenarios(ctx, tenant, formationNames)
 		if err != nil {
 			return err
 		}
@@ -1085,7 +1047,7 @@ func (s *service) ensureApplicationNotPartOfScenarioWithRuntime(ctx context.Cont
 			if err != nil {
 				return errors.Wrapf(err, "while getting application with id %s", appID)
 			}
-			msg := fmt.Sprintf("System %s is still used and cannot be deleted. Unassign the system from the following formations first: %s. Then, unassign the system from the following runtimes, too: %s", application.Name, strings.Join(scenarios, ", "), strings.Join(runtimes, ", "))
+			msg := fmt.Sprintf("System %s is still used and cannot be deleted. Unassign the system from the following formations first: %s. Then, unassign the system from the following runtimes, too: %s", application.Name, strings.Join(formationNames, ", "), strings.Join(runtimes, ", "))
 			return apperrors.NewInvalidOperationError(msg)
 		}
 
@@ -1098,17 +1060,23 @@ func (s *service) ensureApplicationNotPartOfScenarioWithRuntime(ctx context.Cont
 // ensureApplicationNotPartOfAnyScenario Checks if an application has scenarios associated with it. If the application is
 // associated with any scenario it can not be deleted before unassigning from that scenario
 func (s *service) ensureApplicationNotPartOfAnyScenario(ctx context.Context, tenant, appID string) error {
-	scenarios, err := s.getScenarioNamesForApplication(ctx, appID)
+	formations, err := s.formationService.ListFormationsForObject(ctx, appID)
 	if err != nil {
-		return err
+		return errors.Wrapf(err, "while getting formations for Application with ID %s", appID)
 	}
 
-	if len(scenarios) > 0 {
+	if len(formations) > 0 {
 		application, err := s.appRepo.GetByID(ctx, tenant, appID)
 		if err != nil {
 			return errors.Wrapf(err, "while getting application with id %s", appID)
 		}
-		msg := fmt.Sprintf("System %s is part of the following formations : %s", application.Name, strings.Join(scenarios, ", "))
+
+		formationNames := make([]string, 0, len(formations))
+		for _, formation := range formations {
+			formationNames = append(formationNames, formation.Name)
+		}
+
+		msg := fmt.Sprintf("System %s is part of the following formations : %s", application.Name, strings.Join(formationNames, ", "))
 		return apperrors.NewInvalidOperationError(msg)
 	}
 
@@ -1173,6 +1141,10 @@ func (s *service) genericCreate(ctx context.Context, in model.ApplicationRegiste
 		in.Labels[intSysKey] = *in.IntegrationSystemID
 	}
 	in.Labels[nameKey] = normalizedName
+
+	if _, ok := in.Labels[model.ScenariosKey]; ok {
+		return "", errors.Errorf("label with key %s cannot be set explicitly", model.ScenariosKey)
+	}
 
 	err = s.labelService.UpsertMultipleLabels(ctx, appTenant, model.ApplicationLabelableObject, id, in.Labels)
 	if err != nil {
@@ -1279,21 +1251,6 @@ func (s *service) ensureIntSysExists(ctx context.Context, id *string) (bool, err
 	return true, nil
 }
 
-func (s *service) getScenarioNamesForApplication(ctx context.Context, applicationID string) ([]string, error) {
-	log.C(ctx).Infof("Getting scenarios for application with id %s", applicationID)
-
-	scenarios, err := s.GetScenariosGlobal(ctx, applicationID)
-	if err != nil {
-		if apperrors.ErrorCode(err) == apperrors.NotFound {
-			log.C(ctx).Infof("No scenarios found for application")
-			return nil, nil
-		}
-		return nil, err
-	}
-
-	return scenarios, nil
-}
-
 func (s *service) getRuntimeNamesForScenarios(ctx context.Context, tenant string, scenarios []string) ([]string, error) {
 	scenariosQuery := eventing.BuildQueryForScenarios(scenarios)
 	runtimeScenariosFilter := []*labelfilter.LabelFilter{labelfilter.NewForKeyWithQuery(model.ScenariosKey, scenariosQuery)}
@@ -1354,6 +1311,10 @@ func (s *service) genericUpsert(ctx context.Context, appTenant string, in model.
 		in.Labels[intSysKey] = *in.IntegrationSystemID
 	}
 	in.Labels[nameKey] = s.appNameNormalizer.Normalize(app.Name)
+
+	if _, ok := in.Labels[model.ScenariosKey]; ok {
+		return errors.Errorf("label with key %s cannot be set explicitly", model.ScenariosKey)
+	}
 
 	err = s.labelService.UpsertMultipleLabels(ctx, appTenant, model.ApplicationLabelableObject, id, in.Labels)
 	if err != nil {
@@ -1514,18 +1475,6 @@ func createORDWebhookInput(baseURL string, ordWebhookMapping ORDWebhookMapping) 
 			AccessStrategy: str.Ptr(string(accessstrategy.CMPmTLSAccessStrategy)),
 		},
 	}, nil
-}
-
-func createMapFromFormationsSlice(formations []string) map[string]struct{} {
-	resultMap := make(map[string]struct{}, len(formations))
-	for _, f := range formations {
-		resultMap[f] = struct{}{}
-	}
-	return resultMap
-}
-
-func allowAllCriteria(_ string) bool {
-	return true
 }
 
 func buildWebhookProxyURL(mappingCfg ORDWebhookMapping) string {
